Add files via upload
Contrary to the wide held view that more crime occurs in the midnight/ early hours of the morning due to many people being asleep, this analysis has proved that more crime occurs at the late hours of the day. With early hours of the day having the least frequency of crime. The middle of the week to Friday having the highest crime rate. With friday being the day with most crime committed and Subday being the least.

Also, From ending of the 3rd quarter to early months of the 4th quarter being the months with highest month with highest crime rate. September being the month with the most crime rate.

This dataset has helped us gather information about the areas with high rate of crime, and the ability to predict what time and day a crime is likely to occur the highest . These area could be labeled as crime prone areas and other factors that causes the crime can be analyzed to help provide a lasting solution to these areas.

SUMMARY

With this dataset provided, we have been able to identify the most occurring crimes in Boston. These information will help us create more solution to the crime problems' Security agency are to use the above information for planning and deployment.